restructure some folder and change app name on apps tab
