+++ components/Sidebar/Apps/GithubLink.tsx
@@ -4,8 +4,8 @@ import { BrandGithub } from "tabler-icons-react";
 
 export function GithubLink() {
   return (
-    <ListItem title="App Name" description="v.0.1" >
-      <Button component="a" target="_blank" href="https://github.com" variant="outline" leftIcon={<BrandGithub size={14} />}>
+    <ListItem title="Datavizs" description="v.0.1" >
+      <Button component="a" target="_blank" href="https://github.com/jfraziz/datavizs" variant="outline" leftIcon={<BrandGithub size={14} />}>
         Github
       </Button>
     </ListItem>
