+++ components/Sidebar/Data/Import/index.tsx
@@ -1,12 +1,14 @@
 import { useStore } from '@stores/maps';
+import { FileImport } from './FileImport';
 import { useModals } from '@mantine/modals';
 import { Button, Text } from '@mantine/core';
+import { ListItem } from "@components/Sidebar/Common/ListItem";
 import { ModalsContextProps } from '@mantine/modals/lib/context';
-import { GeoJSONFileImport } from '@components/Import/GeoJSONFileImport';
 
-const ImportModalBody: React.FC<{ modals : ModalsContextProps }> = ({ modals }) => <GeoJSONFileImport callback={() => modals.closeAll()} />
 
-export function ImportGeoJSONBtn() {
+const ImportModalBody: React.FC<{ modals: ModalsContextProps }> = ({ modals }) => <FileImport callback={() => modals.closeAll()} />
+
+export function ImportGeoJSON() {
   const modals = useModals();
   const mapKey = useStore.getState().mapKey
 
@@ -32,5 +34,9 @@ export function ImportGeoJSONBtn() {
     });
   }
 
-  return <Button onClick={openContextModal}>Import Data</Button>;
+  return (
+    <ListItem title="Import Data" description="Import geojson">
+      <Button onClick={openContextModal}>Import Data</Button>
+    </ListItem>
+  );
 }
