+++ components/Sidebar/Legends/ColorSwatchs.tsx
