+++ components/Sidebar/Legends/Controls.tsx
