+++ components/Sidebar/Legends/LegendList.tsx
