+++ components/Sidebar/Legends/index.tsx
