+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "maps",
+  "name": "datavizs",
   "version": "0.1.0",
   "private": true,
   "scripts": {
